+++ codes/ConsensusOPLS/NEWS.md
@@ -1,3 +1,22 @@
+## version 1.1.0
+
+---
+
+
+### Changes
+
+- Add predict function
+- Add margin and softmax for confidence score
+- Add interpretation in vignettes
+
+
+### Bugfixes
+
+- Fix koplsCenterK
+- Fix koplsPredict
+- Fix plots
+
+
 ## version 1.0.0
 
 ---